refactor: Refactor which functions are moved from scripts into functions.js
