+++ src/scripts.js
@@ -1,24 +1,11 @@
 // This is the JavaScript entry file - your code begins here
 // Do not delete or rename this file ********
 const {
-  getUserBookings,
-  updateUpcomingVisits,
-  updateSpending,
-  updatePastVisits,
   sortBookingsOldToNew,
   sortBookingsNewToOld,
-  getUpcomingBookings,
-  getPastBookings,
   convertNumToDollarAmount,
   formatDate,
-  makeRoomTypeButton,
-  getBookedRooms,
-  getUnbookedRooms,
-  getAvailableRooms,
-  assignBookedRoom,
-  hide,
-  show,
-} = require('./functions')
+} = require('./functions');
 
 // An example of how you tell webpack to use a CSS (SCSS) file
 import './css/styles.css';
@@ -70,7 +57,7 @@ const availableRoomsContainer = document.querySelector(
   '.available-rooms-container'
 );
 const footer = document.querySelector('.footer');
-const today = new Date();
+let today = new Date();
 let currentUser;
 let allRooms;
 let allBookings;
@@ -141,7 +128,7 @@ const getUser = (url) => {
     .then(() => getAllRooms('http://localhost:3001/api/v1/rooms'))
     .then(() => getAllBookings('http://localhost:3001/api/v1/bookings'))
     .then(() => {
-      getUserBookings();
+      getUserBookings(allBookings, currentUser);
       updateUpcomingVisits();
       populateSpendingAmount();
       updatePastVisits();
@@ -178,6 +165,14 @@ const bookRoom = (url) => {
 };
 
 // Functions
+const hide = (element) => {
+  element.classList.add('hidden');
+};
+
+const show = (element) => {
+  element.classList.remove('hidden');
+};
+
 const changeDashboardView = (event) => {
   if (event.target === loginForm) {
     hide(loginPage);
@@ -327,6 +322,111 @@ const changeDashboardView = (event) => {
     bookRoom('http://localhost:3001/api/v1/bookings');
     getUser(`http://localhost:3001/api/v1/customers/${currentUser.id}`);
   }
+};
+
+const updateUpcomingVisits = () => {
+  const upcomingBookings = getUpcomingBookings(userBookings);
+  const sortedBookings = sortBookingsOldToNew([...upcomingBookings]);
+
+  populateNextVisit(sortedBookings);
+};
+
+const updateSpending = () => {
+  const amount = userBookings.reduce(
+    (acc, booking) => acc + allRooms[booking.roomNumber - 1].costPerNight,
+    0
+  );
+
+  const formattedAmount = convertNumToDollarAmount(amount);
+  return formattedAmount;
+};
+
+const updatePastVisits = () => {
+  const pastBookings = getPastBookings(userBookings);
+  const sortedNewPastBookings = sortBookingsNewToOld([...pastBookings]);
+  const sortedOldPastBookings = sortBookingsOldToNew([...pastBookings]);
+
+  populateRecentVisits(pastBookings);
+};
+
+const getBookedRooms = () => {
+  const checkInDate = checkInDateInput.value
+    .split('Check-in Date')[0]
+    .split('-')
+    .join('/');
+
+  return allBookings
+    .filter((booking) => booking.date === checkInDate)
+    .map((room) => room.roomNumber);
+};
+
+const getUnbookedRooms = (bookedRooms) => {
+  return allRooms.filter((room) => {
+    return !bookedRooms.includes(room.number);
+  });
+};
+
+const getAvailableRooms = () => {
+  const numGuests = numGuestsInput.value.split('num guests')[0];
+  const numBedsNeeded = Math.round(numGuests / 2);
+  const bookedRooms = getBookedRooms();
+  const unbookedRooms = getUnbookedRooms(bookedRooms);
+
+  return unbookedRooms.filter((room) => room.numBeds >= numBedsNeeded);
+};
+
+const assignBookedRoom = () => {
+  const roomContainer = document.querySelector('.individual-room');
+
+  const bedsInfo = document.querySelector('.bed-info').textContent.split(' ');
+  const numBeds = Number(bedsInfo[0]);
+  const bedSize = bedsInfo[1];
+  const costPerNightInfo = document
+    .querySelector('.cost-per-night')
+    .textContent.split(' ')[3];
+  const costPerNight = parseFloat(costPerNightInfo.replace('$', ''));
+
+  bookedRoom = allRooms.filter(
+    (room) =>
+      room.costPerNight === costPerNight &&
+      room.numBeds === numBeds &&
+      room.bedSize === bedSize
+  );
+};
+
+const getUserBookings = (bookings, user) => {
+  userBookings = bookings.filter((booking) => booking.userID === user.id);
+};
+
+const getUpcomingBookings = (bookings) => {
+  return bookings.filter((booking) => {
+    const bookingDate = new Date(booking.date);
+    return bookingDate >= today;
+  });
+};
+
+const getPastBookings = (bookings) => {
+  return bookings.filter((booking) => {
+    const bookingDate = new Date(booking.date);
+    return bookingDate < today;
+  });
+};
+
+const makeRoomTypeButton = (booking) => {
+  if (booking.roomNumber)
+    return allRooms[booking.roomNumber - 1].roomType
+      .split(' ')
+      .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+      .join(' ');
+  else
+    return allRooms[booking.number - 1].roomType
+      .split(' ')
+      .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+      .join(' ');
+};
+
+const formatDateInput = () => {
+  checkInDateInput.min = new Date().toISOString().split('T')[0];
 };
 
 // DOM Manipulation Functions
@@ -598,7 +698,6 @@ const populatePastBookings = () => {
 
   const pastBookingsDiv = document.querySelector('.past-bookings');
 
-  console.log(sortedPastBookings);
   if (sortedPastBookings.length) {
     pastBookingsDiv.innerHTML = '';
 
@@ -753,8 +852,4 @@ const showLoginErrorMessage = (submittedID, submittedPassword) => {
     password.classList.add('sign-in-error');
     password.insertAdjacentElement('afterend', errorText);
   }
-};
-
-const formatDateInput = () => {
-  checkInDateInput.min = new Date().toISOString().split('T')[0];
 };
